BackEnd/python-api/requirements.txt
Flask
Pillow
opencv-python
numpy
gunicorn
